+++ order/src/main/java/com/sparta/hotdeal/order/application/dtos/basket/req/ReqPostBasketDto.java
@@ -1,14 +1,18 @@
-package com.sparta.hotdeal.order.presentation.dtos.basket.req;
+package com.sparta.hotdeal.order.application.dtos.basket.req;
 
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Getter
 @Builder
+@ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReqPatchBasketDto {
+public class ReqPostBasketDto {
+    private UUID privateId;
     private Integer quantity;
 }
